// src_backup/types.ts
export interface Livello {
    id: string;
    nome: string;
    voto: string;
    punteggio: string;
    descrizione: string;
}

export interface Competenza {
    id: string;
    codice: string;
    nome: string;
    framework?: string;
    livelli: Livello[];
    disciplines?: string[];
}
// ...
export interface Rubrica {
    id: string;
    titolo: string;
    criteri: Criterio[];
}

export interface Criterio {
    competenzaId: string;
    indicatori: Indicatore[];
}

export interface Indicatore {
    livelloId: string;
    descrizione: string;
    nota?: string;
}